[DB][BUGFIX] Change DB field sys_filemounts.base from tinyint to integer
sys_filemounts.base holds a reference to a storage, which should
be of DB type "int" as the storage uid is an integer itself as well.

It was a tinyint(4) type and is changed now to int(11).

Resolves: #57584
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/35856
Tested-by: Markus Klein <klein.t3@reelworx.at>

+++ typo3/sysext/core/ext_tables.sql
@@ -241,7 +241,7 @@ CREATE TABLE sys_filemounts (
   tstamp int(11) unsigned DEFAULT '0' NOT NULL,
   title varchar(30) DEFAULT '' NOT NULL,
   path varchar(120) DEFAULT '' NOT NULL,
-  base tinyint(4) unsigned DEFAULT '0' NOT NULL,
+  base int(11) unsigned DEFAULT '0' NOT NULL,
   hidden tinyint(3) unsigned DEFAULT '0' NOT NULL,
   deleted tinyint(1) unsigned DEFAULT '0' NOT NULL,
   sorting int(11) unsigned DEFAULT '0' NOT NULL,
